requirements.txt
faiss-cpu
numpy
streamlit
spacy
sentence-transformers
